// fun-chat/src/types.ts
enum ResponseTypesUsers {
    USER_LOGIN = 'USER_LOGIN',
    USER_LOGOUT = 'USER_LOGOUT',
    USER_EXTERNAL_LOGIN = 'USER_EXTERNAL_LOGIN',
    USER_EXTERNAL_LOGOUT = 'USER_EXTERNAL_LOGOUT',
    USER_ACTIVE = 'USER_ACTIVE',
    USER_INACTIVE = 'USER_INACTIVE',
    ERROR = 'ERROR',
}

enum ResponseTypesChat {
    MSG_SEND = 'MSG_SEND',
    MSG_FROM_USER = 'MSG_FROM_USER',
    MSG_DELIVER = 'MSG_DELIVER',
    MSG_READ = 'MSG_READ',
    MSG_DELETE = 'MSG_DELETE',
    MSG_EDIT = 'MSG_EDIT',
}

export { ResponseTypesUsers, ResponseTypesChat };